requirements.txt
pyjwt
uliweb-alembic
accessify==0.3.1
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
Django==5.2.7
fastapi==0.120.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
jiter==0.9.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==1.78.1
outcome==1.3.0.post0
packaging==25.0
pydantic==2.11.4
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.3
requests==2.32.4
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.48.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
